Add Matlab Menu to Desktop
The trick in this commit is to get SSH with X-Forwarding, without
requiring the user to type anything in the terminal.  To do this, we
must execute the script within a terminal (xterm), and use sshpass.

For password authentication, we use the sshpass recommended method of
storing the password (which we request with Zenity) in a local variable.
We then ssh with all the special options and run Matlab inside of bash.
It works.

If a user enters their password wrong, the terminal will close, and it
will fail silently.